Maps

Description

This module provides an implementation of maps.

Documentation

empty :: Map k v

Empty map, that is, a map containing no bindings.

get :: Eq k => Map k v -> k -> Maybe v

Get a binding from a map.

set :: Map k v -> k -> v -> Map k v

Add a binding to the given map

keys :: Eq k => Map k v -> [k]

Return all the keys in a given map

fromList :: [(k, v)] -> Map k v

Construct a map from the given list of key-value pairs

toList :: Map k v -> [(k, v)]

Return a list of all key-value pairs

data Map k v