Result

Description

This module provides a type of results: either we have a value or we have a failure with an error message. The type is a monad. Several helper functions are provided.

Documentation

data Result a

Represents a successful result with a value or a failuer with an error message.

Constructors

Success a

Failure String

fromMaybe :: Maybe a -> Result a

Convert from a Maybe to a Result, with a default error message for Nothing

fromMaybe' :: String -> Maybe a -> Result a

Convert from a Maybe to a Result, using the given error message for Nothing

toMaybe :: Result a -> Maybe a

Convert a Result to a Maybe

toIO :: Result a -> IO a

Convert a Result to IO