SimpleTests

Description

Documentation

test :: (Eq a, Show a) => String -> a -> a -> IO ()

Test whether the actual value matches the expected one, printing diagnostics.

testError :: Show a => String -> a -> IO ()

Test test for an expected error.

testErrorMsg :: Show a => String -> a -> String -> IO ()

Test test for an expected error with a specific message.

testStats :: IO ()

Print the test stats.

beginTests :: IO ()

Set up a test batch. Resets test counters to 0.

endTests :: IO ()

End tests, print test stats.

testSection :: String -> IO ()

Label a test section